requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
beautifulsoup4==4.12.3

